+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/section/MacrosSection.java
@@ -107,6 +107,7 @@ public class MacrosSection implements ISectionBuilder {
         final DocSection assert_ = new DocSection("Assert", "macros.assert");
         all.addSection(assert_);
         assert_.addItem(diBuilder.getDocItem("assert", true, true));
+        assert_.addItem(diBuilder.getDocItem("assert-throws", true, true));
 
         final DocSection util = new DocSection("Util", "macros.util");
         all.addSection(util);