+++ ChangeLog.md
@@ -14,7 +14,8 @@ and this project adheres to
 
 ### Added
 
-- support for destructuring for `loop / recur` loops
+- support for destructuring in `loop / recur` loops
+- macro `assert-throws`
 - an optional customized indentation string to the JSON pretty printer
 - proxy builder for Java functional UnaryPredicate interface